New Changes

+++ Course/execises/medidas.c
@@ -1,30 +1,25 @@
 #include <stdio.h>
 #include <math.h>
-#include <stdlib.h>
-#include <locale.h>
 
 int main() {
 
-    float medidaa, medidab, medidac, quadrado, triangulo, trapezio;
+    double medidaa, medidab, medidac, quadrado, triangulo, trapezio;
 
     printf("Digite a medida A:");
-    scanf("%f", &medidaa);
+    scanf("%lf", &medidaa);
 
     printf("Digite a medida B:");
-    scanf("%f", &medidab);
+    scanf("%lf", &medidab);
 
     printf("Digite a medida C:");
-    scanf("%f", &medidac);
+    scanf("%lf", &medidac);
 
     quadrado = pow(medidaa, 2);
-
     triangulo = (medidaa * medidab) / 2;
-
     trapezio = ((medidaa + medidab) * medidac) / 2;
 
-    printf("AREA DO QUADRADO = %0.4f", quadrado);
-    printf("\nAREA DO TRAINGULO = %0.4f", triangulo);
-    printf("\n AREA DO TRAPEZIO = %0.4f", trapezio);
-
-    
+    printf("AREA DO QUADRADO = %0.4lf", quadrado);
+    printf("\nAREA DO TRAINGULO = %0.4lf", triangulo);
+    printf("\n AREA DO TRAPEZIO = %0.4lf", trapezio);
+   
 }
