+++ Course/execises2/operadora.c
@@ -2,21 +2,21 @@
 
 int main ()
 {
-    float minutos, pagar;
+    double minutos, pagar;
 
     printf("Digite a quantidade de minutos:");
-    scanf("%f", &minutos);
+    scanf("%lf", &minutos);
 
-    if  ( minutos < 100)
+    if (minutos < 100)
     {
         pagar = 50;    
     }
     
-    else {
-
-    pagar = ((minutos - 100) * 2) + 50; 
+    else 
+    {
+        pagar = ((minutos - 100) * 2) + 50; 
 
     }
     
-    printf("Valor a pagar: %.2f ", pagar);
+    printf("Valor a pagar: %.2lf ", pagar);
 }
